Refactor files/images.py into separate images/*.py files
